(TWILL-125) Eliminate mainObject instance in BundedJarRunner
Instantiating the object containing the main method in order to invoke
the main method is unnecessary since main is always static. There is
also a potential problem with calling newInstance() on the main object's
class; it will throw an exception if the main object does not have
a 0-args constructor.

Signed-off-by: Terence Yim <chtyim@apache.org>