TWILL-128 Change the ZKBrokerService to be public so it could be used outside the package.
Make ZKBrokerService to have public modifier so it could be used outside the package private.

The BrokerService interface that ZKBrokerService implemented is already public so it would be
helpful to make this implementation to be public so it could be used by external projects or code
when using Apache Kafka with Simple Customer APIs to resolve leader.

Signed-off-by: Terence Yim <chtyim@apache.org>