TWILL-24 Add log level settable for Twill application.
This PR allow client to set log level for the Twill applications in the container level.

The TwillPreparer now has new TwillPreparer#enableLogLevel method to pass log level that
will be set on env variable in app master and pass to app containers.

Add unit test to test the new feature

This closes #39 on GitHub

Signed-off-by: Terence Yim <chtyim@apache.org>