TWILL-137 Be able to get log level for each runnable via TwillController.getResourceReport
With TWILL-24 resolved, the client should be able to get the log level for an application
for each runnable.

This involves adding log level to TwillRunResources interface and its implementations.

Changes include adding log level to the TwillRunResources and modify unit tests to test
the new property.

This closes #46 on GitHub

Signed-off-by: Terence Yim <chtyim@apache.org>